Move CMPSPInfo into seperate file
